removes resource rename via AST in dlt init, provides new templates
